Converted static strings to F() to use flash instead of sram, added sram free memory check, defined periods for diagnostics timers, added in extra pins for quad encoder and interrupts, formatiing changes

+++ Arduino-Code/main/realTimer.h
@@ -1,5 +1,6 @@
 //Creates a timer that operates on a real time check based on the arduino internal clock
 //Used as an alarm or to create time based loops
+//Uses milliseconds
 //Could have over flow issues if runtime is too long
 
 #ifndef REAL_TIMER_H
@@ -7,12 +8,18 @@
 
 class realTimer{
   private:
-    double accum;
+    //Private variables
+    unsigned long accum;
     double lastCheck;
     double enable;
-    double targetTime;
+    unsigned long targetTime;
+
+    //Private variables
     
   public:
+    //Public variables
+    
+    //Public functions
     void init(double setTime);
     bool check(bool condition);
 };
